Introducing "CPU sets" (std::vector<bool>) to identify available cpus and affinity masks.  And test thread affinities.
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@979 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ lass/util/bit_manip.h
@@ -81,14 +81,14 @@ namespace lass
 namespace util
 {
 
-template<typename T> inline void setBit(T& a_bits, unsigned a_bit);
-template<typename T> inline void clearBit(T& a_bits, unsigned a_bit);
-template<typename T> inline void flipBit(T& a_bits, unsigned a_bit);
-template<typename T> inline void setBitIf(T& a_bits, unsigned a_bit, bool a_condition);
-template<typename T> inline void clearBitIf(T& a_bits, unsigned a_bit, bool a_condition);
-template<typename T> inline void flipBitIf(T& a_bits, unsigned a_bit, bool a_condition);
-template<typename T> inline void setBitTo(T& a_bits, unsigned a_bit, bool a_state);
-template<typename T> inline bool checkBit(T a_bits, unsigned a_bit);
+template<typename T> inline void setBit(T& a_bits, size_t a_bit);
+template<typename T> inline void clearBit(T& a_bits, size_t a_bit);
+template<typename T> inline void flipBit(T& a_bits, size_t a_bit);
+template<typename T> inline void setBitIf(T& a_bits, size_t a_bit, bool a_condition);
+template<typename T> inline void clearBitIf(T& a_bits, size_t a_bit, bool a_condition);
+template<typename T> inline void flipBitIf(T& a_bits, size_t a_bit, bool a_condition);
+template<typename T> inline void setBitTo(T& a_bits, size_t a_bit, bool a_state);
+template<typename T> inline bool checkBit(T a_bits, size_t a_bit);
 
 template<typename T> inline void setMasked(T& a_bits, const T& a_mask);
 template<typename T> inline void clearMasked(T& a_bits, const T& a_mask);
